Support initializing Hwacha with a configuration block
The simple integer argument for max concurrency is ok for basic use, but
adding on additional arguments to support additional configuration is
not ideal.  Better to yield a data object to a configuration block. That
means adding on new configuration options is easy and code using Hwacha
is easier to follow.

Hwacha.new(10) conveys nothing about what 10 is actually doing. A block
for configuration is much more self-documenting.

Hwacha.new do |config|
  config.max_concurrent_requests = 10
end