Replace VCR with Typhoeus::Response stubs
VCR is great for hitting an actual external API and replaying its
responses to be sure that your code is handling them correctly. But
Hwacha doesn't really care about the content, only that it's dealing
with Typhoeus responses correctly. That means Typhoeus's built-in
support for stubbing responses is a perfect fit for Hwacha testing.