Support following redirects
This was a bit more of a change that it seemed that it would be. The
"followlocation" option needs to be set on Typhoeus::Request not
Typhoeus::Hydra. That meant that Hwacha::Config needed to have two
different configuration exports: hydra_options and request_options.

Stubbing out Typhoeus responses that properly simulate a redirect chain
would be very complicated. And arguably the methods 'check' and
'find_existing' only care about properly dealing with the final HTTP
state of success or failure. So testing the proper handling of redirects
is left to the integration spec.